♻️ refactor: 상단 동아리 정보 컴포넌트 분리
