+++ src/components/applicant/application_history/detail/container/RecrutingInfoContainer/Chart.tsx
