+++ src/components/applicant/application_history/detail/container/RecrutingInfoContainer/RecrutingInfoContainer.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Chart from "./Chart";
 
 export default function RecrutingInfoContainer() {
